+++ openhands-agent-server/openhands/agent_server/models.py
@@ -31,10 +31,6 @@ from openhands.sdk.llm.utils.metrics import MetricsSnapshot
 from openhands.sdk.profiles.agent_profile import (
     LaunchedAgentProfile as LaunchedAgentProfile,
 )
-from openhands.sdk.runtime_context import (
-    ConversationRuntimeContext as ConversationRuntimeContext,
-    RuntimeService as RuntimeService,
-)
 from openhands.sdk.secret import SecretSource
 from openhands.sdk.security.analyzer import SecurityAnalyzerBase
 from openhands.sdk.security.confirmation_policy import (
@@ -88,6 +84,7 @@ class StoredConversation(StartConversationRequest):
     # agent_profile_id is resolved into launched_agent_profile at creation; exclude from
     # the persistence payload so it does not re-appear in meta.json.
     agent_profile_id: UUID | None = Field(default=None, exclude=True)
+
     id: OpenHandsUUID
     title: str | None = Field(
         default=None, description="User-defined title for the conversation"
@@ -136,10 +133,6 @@ class _ConversationInfoBase(BaseModel):
         default="workspace/conversations",
         description="Directory for persisting conversation state and events. "
         "If None, conversation will not be persisted.",
-    )
-    runtime_context: ConversationRuntimeContext | None = Field(
-        default=None,
-        description="Deployment and service topology for this conversation.",
     )
     max_iterations: int = Field(
         default=500,
