+++ openhands-sdk/openhands/sdk/__init__.py
@@ -16,10 +16,8 @@ from openhands.sdk.conversation import (
     Conversation,
     ConversationCallbackType,
     ConversationExecutionStatus,
-    ConversationRuntimeContext,
     LocalConversation,
     RemoteConversation,
-    RuntimeService,
 )
 from openhands.sdk.conversation.conversation_stats import ConversationStats
 from openhands.sdk.event import Event, HookExecutionEvent, LLMConvertibleEvent
@@ -154,9 +152,7 @@ __all__ = [
     "LocalConversation",
     "RemoteConversation",
     "ConversationExecutionStatus",
-    "ConversationRuntimeContext",
     "ConversationCallbackType",
-    "RuntimeService",
     "Event",
     "LLMConvertibleEvent",
     "AgentContext",
