+++ openhands-sdk/openhands/sdk/context/prompts/presets.py
@@ -21,7 +21,6 @@ from openhands.sdk.context.prompts.sections.dynamic import (
     CustomSuffixSection,
     DateTimeSection,
     RepoContextSection,
-    RuntimeServicesSection,
 )
 from openhands.sdk.context.prompts.sections.planning import PlanningSection
 from openhands.sdk.context.prompts.sections.static import (
@@ -83,7 +82,6 @@ _DYNAMIC_SECTIONS: Final[tuple[PromptSection, ...]] = (
     RepoContextSection(),  # guard: gated repo skills present
     AvailableSkillsSection(),  # guard: available_skills_prompt
     CustomSuffixSection(),  # guard: system_message_suffix
-    RuntimeServicesSection(),  # guard: conversation runtime services
     CustomSecretsSection(),  # guard: secret_infos present
     # DateTimeSection is intentionally last: it is the only per-conversation
     # volatile value, so the stable dynamic content stays a cache-friendly prefix.
