+++ openhands-sdk/openhands/sdk/context/prompts/sections/__init__.py
@@ -13,7 +13,6 @@ from openhands.sdk.context.prompts.sections.dynamic import (
     CustomSuffixSection,
     DateTimeSection,
     RepoContextSection,
-    RuntimeServicesSection,
 )
 from openhands.sdk.context.prompts.sections.planning import PlanningSection
 from openhands.sdk.context.prompts.sections.static import (
@@ -56,7 +55,6 @@ __all__ = [
     "ProcessManagementSection",
     "PullRequestsSection",
     "RepoContextSection",
-    "RuntimeServicesSection",
     "RoleSection",
     "SecurityRiskAssessmentSection",
     "SecuritySection",
