+++ openhands-sdk/openhands/sdk/conversation/__init__.py
@@ -26,10 +26,6 @@ from openhands.sdk.conversation.visualizer import (
     ConversationVisualizerBase,
     DefaultConversationVisualizer,
 )
-from openhands.sdk.runtime_context import (
-    ConversationRuntimeContext,
-    RuntimeService,
-)
 
 
 __all__ = [
@@ -38,7 +34,6 @@ __all__ = [
     "BaseConversation",
     "ConversationState",
     "ConversationExecutionStatus",
-    "ConversationRuntimeContext",
     "ConversationCallbackType",
     "ConversationTags",
     "ConversationTokenCallbackType",
@@ -49,7 +44,6 @@ __all__ = [
     "EventLog",
     "ResourceLockManager",
     "ResourceLockTimeout",
-    "RuntimeService",
     "LocalConversation",
     "RemoteConversation",
     "EventsListBase",
