+++ openhands-sdk/openhands/sdk/conversation/conversation.py
@@ -18,7 +18,6 @@ from openhands.sdk.conversation.visualizer import (
 from openhands.sdk.hooks import HookConfig
 from openhands.sdk.logger import get_logger
 from openhands.sdk.plugin import PluginSource
-from openhands.sdk.runtime_context import ConversationRuntimeContext
 from openhands.sdk.secret import SecretValue
 from openhands.sdk.tool.client_tool import ClientToolSpec
 from openhands.sdk.utils.redact import redact_url_credentials
@@ -89,7 +88,6 @@ class Conversation:
         observability_metadata: dict[str, TraceMetadataValue] | None = None,
         observability_tags: list[str] | None = None,
         observability_span_name: str = "conversation",
-        runtime_context: ConversationRuntimeContext | None = None,
     ) -> "LocalConversation": ...
 
     @overload
@@ -119,7 +117,6 @@ class Conversation:
         observability_metadata: dict[str, TraceMetadataValue] | None = None,
         observability_tags: list[str] | None = None,
         observability_span_name: str = "conversation",
-        runtime_context: ConversationRuntimeContext | None = None,
     ) -> "RemoteConversation": ...
 
     def __new__(
@@ -149,7 +146,6 @@ class Conversation:
         observability_metadata: dict[str, TraceMetadataValue] | None = None,
         observability_tags: list[str] | None = None,
         observability_span_name: str = "conversation",
-        runtime_context: ConversationRuntimeContext | None = None,
     ) -> BaseConversation:
         from openhands.sdk.conversation.impl.local_conversation import LocalConversation
         from openhands.sdk.conversation.impl.remote_conversation import (
@@ -210,7 +206,6 @@ class Conversation:
                 tags=effective_tags if effective_tags else None,
                 user_id=user_id,
                 client_tools=client_tools,
-                runtime_context=runtime_context,
                 observability_metadata=observability_metadata,
                 observability_tags=observability_tags,
                 observability_span_name=observability_span_name,
@@ -234,7 +229,6 @@ class Conversation:
             tags=tags,
             user_id=user_id,
             client_tools=client_tools,
-            runtime_context=runtime_context,
             observability_metadata=observability_metadata,
             observability_tags=observability_tags,
             observability_span_name=observability_span_name,
