+++ openhands-sdk/openhands/sdk/conversation/request.py
@@ -8,15 +8,18 @@ agent-server.
 
 from __future__ import annotations
 
+import re
 from typing import Annotated, Any, Literal, cast
 from uuid import UUID
 
 from pydantic import (
     BaseModel,
+    ConfigDict,
     Discriminator,
     Field,
     Tag,
     field_serializer,
+    field_validator,
     model_validator,
 )
 
@@ -32,7 +35,6 @@ from openhands.sdk.conversation.types import (
 from openhands.sdk.hooks import HookConfig
 from openhands.sdk.llm.message import ImageContent, Message, TextContent
 from openhands.sdk.plugin import PluginSource
-from openhands.sdk.runtime_context import ConversationRuntimeContext
 from openhands.sdk.secret import SecretSource
 from openhands.sdk.security.analyzer import SecurityAnalyzerBase
 from openhands.sdk.security.confirmation_policy import (
@@ -73,6 +75,92 @@ class SendMessageRequest(BaseModel):
 
     def create_message(self) -> Message:
         return Message(role=self.role, content=self.content)
+
+
+_ENV_NAME = re.compile(r"^[A-Za-z_][A-Za-z0-9_]*$")
+_HEADER_NAME = re.compile(r"^[!#$%&'*+.^_`|~0-9A-Za-z-]+$")
+_SERVICE_NAME = re.compile(r"^[a-z][a-z0-9_-]*$")
+_MODE_NAME = re.compile(r"^[A-Za-z0-9][A-Za-z0-9._:-]*$")
+
+
+class RuntimeService(BaseModel):
+    """Describe a service reachable from the agent runtime."""
+
+    model_config = ConfigDict(extra="forbid")
+
+    name: str = Field(min_length=1, max_length=64)
+    url_from_agent: str | None = Field(default=None, max_length=2048)
+    api_prefix: str | None = Field(default=None, max_length=512)
+    docs_url: str | None = Field(default=None, max_length=2048)
+    openapi_url: str | None = Field(default=None, max_length=2048)
+    auth_header_name: str | None = Field(default=None, max_length=128)
+    auth_env_var: str | None = Field(default=None, max_length=128)
+    available: bool = True
+
+    @field_validator("name")
+    @classmethod
+    def _validate_name(cls, value: str) -> str:
+        if not _SERVICE_NAME.fullmatch(value):
+            raise ValueError("name must be a lowercase service identifier")
+        return value
+
+    @field_validator("url_from_agent", "docs_url", "openapi_url")
+    @classmethod
+    def _validate_url(cls, value: str | None) -> str | None:
+        if value is not None and (
+            any(char.isspace() for char in value) or "<" in value or ">" in value
+        ):
+            raise ValueError("runtime service URLs must be single-line values")
+        return value
+
+    @field_validator("api_prefix")
+    @classmethod
+    def _validate_api_prefix(cls, value: str | None) -> str | None:
+        if value is not None and (
+            not value.startswith("/") or any(char.isspace() for char in value)
+        ):
+            raise ValueError("api_prefix must be a single-line absolute path")
+        return value
+
+    @field_validator("auth_header_name")
+    @classmethod
+    def _validate_auth_header_name(cls, value: str | None) -> str | None:
+        if value is not None and not _HEADER_NAME.fullmatch(value):
+            raise ValueError("auth_header_name must be a valid HTTP header name")
+        return value
+
+    @field_validator("auth_env_var")
+    @classmethod
+    def _validate_auth_env_var(cls, value: str | None) -> str | None:
+        if value is not None and not _ENV_NAME.fullmatch(value):
+            raise ValueError("auth_env_var must be a valid environment variable name")
+        return value
+
+
+class RuntimeServices(BaseModel):
+    """Describe the services bound to one conversation launch."""
+
+    model_config = ConfigDict(extra="forbid")
+
+    mode: str | None = Field(default=None, max_length=64)
+    services: list[RuntimeService] = Field(min_length=1)
+
+    @field_validator("mode")
+    @classmethod
+    def _validate_mode(cls, value: str | None) -> str | None:
+        if value is not None and not _MODE_NAME.fullmatch(value):
+            raise ValueError("mode must be a deployment identifier")
+        return value
+
+    @field_validator("services")
+    @classmethod
+    def _validate_unique_services(
+        cls, value: list[RuntimeService]
+    ) -> list[RuntimeService]:
+        names = [service.name for service in value]
+        if len(names) != len(set(names)):
+            raise ValueError("runtime service names must be unique")
+        return value
 
 
 class StartConversationRequest(BaseModel):
@@ -162,6 +250,13 @@ class StartConversationRequest(BaseModel):
             "observation immediately."
         ),
     )
+    runtime_services: RuntimeServices | None = Field(
+        default=None,
+        description=(
+            "Deployment services to add to the resolved agent for this launch. "
+            "The stored Agent Profile is not modified."
+        ),
+    )
     agent_definitions: list[AgentDefinition] = Field(
         default_factory=list,
         description=(
@@ -239,13 +334,6 @@ class StartConversationRequest(BaseModel):
             "model for titles regardless of the agent's main model. If not "
             "set (or profile loading fails), title generation falls back to "
             "the agent's LLM."
-        ),
-    )
-    runtime_context: ConversationRuntimeContext | None = Field(
-        default=None,
-        description=(
-            "Deployment and service topology for this conversation. Stored as "
-            "conversation state and rendered separately from Agent configuration."
         ),
     )
